% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{cellIntensity<-}
\alias{cellIntensity<-}
\title{sets cell Intensity slot to a new matrix. rows protein, columns are cells.}
\usage{
cellIntensity(object) <- value
}
\description{
sets cell Intensity slot to a new matrix. rows protein, columns are cells.
}
