% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{cellIntensity}
\alias{cellIntensity}
\alias{cellIntensity,imcExperiment-method}
\alias{cellIntensity<-}
\alias{cellIntensity<-,imcExperiment,matrix-method}
\title{finds the intensities getter.}
\usage{
cellIntensity(object, ...)

\S4method{cellIntensity}{imcExperiment}(object)

cellIntensity(object) <- value

\S4method{cellIntensity}{imcExperiment,matrix}(object) <- value
}
\arguments{
\item{object}{IMC container}

\item{...}{additional arguments}

\item{value}{matrix rows protein, columns are cells}
}
\description{
finds the intensities getter.

sets cell Intensity slot to a new matrix. rows protein, columns are cells.
}
