% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-class.R
\name{.checkSpatialDimension}
\alias{.checkSpatialDimension}
\title{the rows are the panel names, the columns are the single cells,the column are the single cells to match the SCE designs (scRNA)}
\usage{
.checkSpatialDimension(object)
}
\arguments{
\item{object}{imcExperiment object, class imcExperiment container}
}
\value{
imcExperiment container that has proper dimensions
}
\description{
the rows are the panel names, the columns are the single cells,the column are the single cells to match the SCE designs (scRNA)
}
