% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperimentToHyperFrame.R
\name{.imcExperimentToPPP}
\alias{.imcExperimentToPPP}
\title{map to point pattern from imcExperiment class.}
\usage{
.imcExperimentToPPP(caseExperiment = NULL, phenotypeToUse = 1)
}
\description{
map to point pattern from imcExperiment class.
}
