% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{getCoordinates<-}
\alias{getCoordinates<-}
\title{Sets the coordinate positions of each cell (matrix), columns are X,Y positions.}
\usage{
getCoordinates(object) <- value
}
\arguments{
\item{object}{is IMC container}

\item{value}{matrix rows cells, columns are x,y}
}
\value{
imcExperiment container
}
\description{
Sets the coordinate positions of each cell (matrix), columns are X,Y positions.
}
\examples{
data(imcdata)
x<-getCoordinates(imcdata)
getCoordinates(imcdata)<-as.matrix(x)
}
