% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{getDistance<-}
\alias{getDistance<-}
\title{re-assigns the distance matrix (rows are cells)}
\usage{
getDistance(object) <- value
}
\arguments{
\item{object}{is IMC container}

\item{value}{matrix rows cells, columns are distance measurements}
}
\value{
imcExperiment container
}
\description{
re-assigns the distance matrix (rows are cells)
}
\examples{
data(imcdata)
newD<-matrix(1,nrow=ncol(imcdata),ncol=1)
getDistance(imcdata)<-newD
}
