% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{getMorphology<-}
\alias{getMorphology<-}
\title{re-assigns morphological features can be stored (matrix) rows are cells and columns are Area, etc.}
\usage{
getMorphology(object) <- value
}
\arguments{
\item{object}{is IMC container}

\item{value}{matrix rows cells, columns are Area, Eccentricity, etc.}
}
\description{
re-assigns morphological features can be stored (matrix) rows are cells and columns are Area, etc.
}
