% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{getNeighborhood<-}
\alias{getNeighborhood<-}
\title{slow assignment for the histoCAT neighborhood data (matrix) columns are the neighbors}
\usage{
getNeighborhood(object) <- value
}
\arguments{
\item{object}{is IMC container}

\item{value}{matrix rows cells, columns are neighborhood histocat output}
}
\description{
slow assignment for the histoCAT neighborhood data (matrix) columns are the neighbors
}
