% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{getNetwork<-}
\alias{getNetwork<-}
\title{re-assigns the network assignment (matrix)}
\usage{
getNetwork(object) <- value
}
\arguments{
\item{object}{is IMC container}

\item{value}{matrix rows cells, columns are phenograph network ID}
}
\description{
re-assigns the network assignment (matrix)
}
