% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcData.R
\docType{data}
\name{imcData}
\alias{imcData}
\title{histoCAT output containerized as IMC container.
IMC S4 data set containing 2,452 cells and 44 antibody features which include panel antibody.}
\format{
A data frame of 2,452 cells and histoCAT features
}
\source{
\url{www.usc.edu/}
}
\usage{
imcData
}
\description{
histoCAT output containerized as IMC container.
IMC S4 data set containing 2,452 cells and 44 antibody features which include panel antibody.
}
\keyword{datasets}
