% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-class.R,
%   R/imcExperiment-methods.R
\docType{class}
\name{imcExperiment-class}
\alias{imcExperiment-class}
\alias{.imcExperiment}
\alias{getNetwork,imcExperiment-method}
\alias{getNetwork}
\alias{imcExperiment-method}
\alias{getNetwork<-,imcExperiment,data.frame-method}
\alias{getDistance,imcExperiment-method}
\alias{getDistance}
\alias{getDistance<-,imcExperiment,matrix-method}
\alias{getMorphology,imcExperiment-method}
\alias{getMorphology}
\alias{getMorphology<-,imcExperiment,matrix-method}
\alias{spatial}
\alias{getLabel,imcExperiment-method}
\alias{getLabel}
\title{a summarized experiment of IMC runs, dimensions of the spatial and intensity data are regulated.#'}
\usage{
getNetwork(object)

\S4method{getNetwork}{imcExperiment}(object)

\S4method{getNetwork}{imcExperiment,data.frame}(object) <- value

getDistance(object)

\S4method{getDistance}{imcExperiment}(object)

\S4method{getDistance}{imcExperiment,matrix}(object) <- value

getMorphology(object)

\S4method{getMorphology}{imcExperiment}(object)

\S4method{getMorphology}{imcExperiment,matrix}(object) <- value

getLabel(object)

\S4method{getLabel}{imcExperiment}(object)
}
\arguments{
\item{object}{imcExperiment}

\item{value}{matrix rows cells, columns are Area, etc.}
}
\description{
a summarized experiment of IMC runs, dimensions of the spatial and intensity data are regulated.#'

finds the network information.

assigns cell cluster assignment to the container. rows are cells and column is the cluster ID

finds the distance information.

distance matrix can be stored in the distance slot for pairwise distance

finds the morphology information.

morphological features can be stored (matrix) rows are cells and columns are Area, etc.

finds the label information.

unique cell labels can be assigned (vector)
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{Spatial coordinate (rows are cells and columns are x,y coordinates.}

\item{\code{cellIntensity}}{a matrix of the intensity values rows are proteins and columns are cells (SCE format)}

\item{\code{neighborHood}}{a matrix (rows are cells, columns are features) of neighborhood observations from neighborhood analysis}

\item{\code{network}}{a matrix (rows are cells) and columns are for network assignment information}

\item{\code{distance}}{a matrix of rows are cells and columns are for distance measurements}

\item{\code{morphology}}{matrix of morphological features}

\item{\code{uniqueLabel}}{a character for each cell for unique labeling.}
}}

\section{}{
NA
}

