% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment.R
\name{imcExperiment}
\alias{imcExperiment}
\title{Initializes a imcExperiment and performs some rudimentary checks.
Many of the arguments CAN be NULL; determination of which is required 
is done at run-time.  A imcExperiment must contain at least the 
expressions and spatial assays.}
\usage{
imcExperiment(spatial = matrix(1, 3, 3), cellIntensity = matrix(1, 3,
  3), neighborHood = matrix(1, 3, 3), network = matrix(1, 3, 3),
  uniqueLabel = rep("A", 3), ROIID = data.frame(ROIID = rep("A", 3)),
  ...)
}
\arguments{
\item{spatial}{coordinate data}

\item{exprs_intensity}{matrix of counts}

\item{covariates}{the column metadata (covariates) for each sample}
}
\description{
FIXME: add Harold's quasi-normalized TPM method for tx-level comparisons.
}
