% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperimentToHyperFrame.R
\name{imcExperimentToHyperFrame}
\alias{imcExperimentToHyperFrame}
\title{map to point pattern from imcExperiment class.}
\usage{
imcExperimentToHyperFrame(imcExperiment = NULL)
}
\arguments{
\item{imcExperiment}{imcExperiment class}
}
\description{
map to point pattern from imcExperiment class.
}
