% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentilenormalize.R
\name{percentilenormalize}
\alias{percentilenormalize}
\title{given a  matrix of intensity counts, perform min/max norm.}
\usage{
percentilenormalize(data = NULL, percentile = NULL)
}
\arguments{
\item{data}{matrix of numeric data only}

\item{percentile}{numeric value 0.99 default.}
}
\value{
normalized data, each column on [0,1] scale.
}
\description{
given a  matrix of intensity counts, perform min/max norm.
}
\examples{
 data(data)
 dim(data)
 expr<-data[,3:36]
 normExp<-percentilenormalize(data=expr,percentile=0.99)
 normExp<-as.matrix(normExp)
}
