% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustering_heatmap_wrapper.R
\name{plot_clustering_heatmap_wrapper}
\alias{plot_clustering_heatmap_wrapper}
\title{plot the heatmap of phenotypes using imcExperiment container.}
\usage{
plot_clustering_heatmap_wrapper(myExperiment = NULL, useMedians = TRUE,
  color_clusters = NULL)
}
\arguments{
\item{myExperiment}{imcExperiment}
}
\value{
a GRanges or a KallistoExperiment, depending on `what`
}
\description{
plot the heatmap of phenotypes using imcExperiment container.
}
