% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.R
\docType{pre}
\name{pre}
\alias{pre}
\title{pre}
\format{
A data frame of 180539 cells and histoCAT features
\describe{
  \item{ImageId}{feature from histoCAT}
  \item{CellId}{feature from histoCAT}
 \item{Area}{feature from histoCAT}
 \item{Eccentricity}{feature from histoCAT}
 \item{Solidity}{feature from histoCAT}
 \item{Extent}{feature from histoCAT}
 \item{EulerNumber}{feature from histoCAT}
 \item{Perimeter}{feature from histoCAT}
 \item{MajorAxisLength}{feature from histoCAT}
 \item{MinorAxisLength}{feature from histoCAT}
 \item{Orientation}{feature from histoCAT}
 \item{X_position}{feature from histoCAT}
 \item{Y_position}{feature from histoCAT}
 \item{Percent_Touching}{feature from histoCAT}
 \item{Number_Neighbors}{feature from histoCAT}
 \item{neighbour_4_CellId1}{feature from histoCAT}
 \item{neighbour_4_CellId2}{feature from histoCAT}
 \item{neighbour_4_CellId3}{feature from histoCAT}
 \item{neighbour_4_CellId4}{feature from histoCAT}
 \item{neighbour_4_CellId5}{feature from histoCAT}
 \item{ROIID}{feature from histoCAT}
 \item{uniqueLabel}{feature from histoCAT}
 \item{Patient.Number}{feature from histoCAT}
 \item{condition}{feature from histoCAT}
 \item{Pano}{feature from histoCAT}
 \item{ROI}{feature from histoCAT}
 \item{Tissue.Type}{feature from histoCAT}
 \item{Age}{feature from histoCAT}
 \item{Cytogenetics}{feature from histoCAT}
 \item{TMA}{feature from histoCAT}
 \item{batch}{feature from histoCAT}
 \item{ROIID.1}{feature from histoCAT}
   }
}
\usage{
data(pre)
}
\description{
Data set containing 180539 cells and 32 features simulated from CellProfiler output which include  neighborhood computations, and phenograph clustering.
}
\keyword{datasets}
