% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{selectCases}
\alias{selectCases}
\alias{selectCases,imcExperiment-method}
\title{subsets the imcExperiment to a case along with all slots for a selected multiple ROIs.}
\usage{
selectCases(object, value, ...)

\S4method{selectCases}{imcExperiment}(object, value)
}
\arguments{
\item{object}{IMC container}

\item{value}{this is ROIID vector}

\item{...}{additional parameters}
}
\description{
subsets the imcExperiment to a case along with all slots for a selected multiple ROIs.

method to subset the slots, requires colData with column "ROIID"
}
