% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{histoCAT output saved as a data frame which can be set into S4 class.
Data set containing 1,000 cells and 73 features which include panel antibody, neighborhood computations, and phenograph clustering.}
\format{
A data frame of 1,000 cells and histoCAT features
}
\usage{
data
}
\description{
histoCAT output saved as a data frame which can be set into S4 class.
Data set containing 1,000 cells and 73 features which include panel antibody, neighborhood computations, and phenograph clustering.
}
\keyword{datasets}
