% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{data}
\format{
A data frame of 1,000 cells and histoCAT features
\describe{
  \item{ImageId}{feature from histoCAT}
  \item{CellId}{feature from histoCAT}
  \item{marker1}{feature from histoCAT}
  \item{marker2}{feature from histoCAT}
  \item{marker3}{feature from histoCAT}
  \item{marker4}{feature from histoCAT}
  \item{marker5}{feature from histoCAT}
  \item{marker6}{feature from histoCAT}
  \item{marker7}{feature from histoCAT}
  \item{marker8}{feature from histoCAT}
  \item{marker9}{feature from histoCAT}
  \item{marker10}{feature from histoCAT}
  \item{marker11}{feature from histoCAT}
  \item{marker12}{feature from histoCAT}
  \item{marker13}{feature from histoCAT}
  \item{marker14}{feature from histoCAT}
  \item{marker15}{feature from histoCAT}
  \item{marker16}{feature from histoCAT}
  \item{marker17}{feature from histoCAT}
  \item{marker18}{feature from histoCAT}
  \item{marker19}{feature from histoCAT}
  \item{marker20}{feature from histoCAT}
  \item{marker21}{feature from histoCAT}
  \item{marker22}{feature from histoCAT}
  \item{marker23}{feature from histoCAT}
  \item{marker24}{feature from histoCAT}
  \item{marker25}{feature from histoCAT}
  \item{marker26}{feature from histoCAT}
  \item{marker27}{feature from histoCAT}
  \item{marker28}{feature from histoCAT}
  \item{marker29}{feature from histoCAT}
  \item{marker30}{feature from histoCAT}
  \item{marker31}{feature from histoCAT}
  \item{marker32}{feature from histoCAT}
  \item{marker33}{feature from histoCAT}
  \item{marker34}{feature from histoCAT}
 \item{Area}{feature from histoCAT}
 \item{Eccentricity}{feature from histoCAT}
 \item{Solidity}{feature from histoCAT}
 \item{Extent}{feature from histoCAT}
 \item{EulerNumber}{feature from histoCAT}
 \item{Perimeter}{feature from histoCAT}
 \item{MajorAxisLength}{feature from histoCAT}
 \item{MinorAxisLength}{feature from histoCAT}
 \item{Orientation}{feature from histoCAT}
 \item{X_position}{feature from histoCAT}
 \item{Y_position}{feature from histoCAT}
 \item{Percent_Touching}{feature from histoCAT}
 \item{Number_Neighbors}{feature from histoCAT}
 \item{neighbour_4_CellId1}{feature from histoCAT}
 \item{neighbour_4_CellId2}{feature from histoCAT}
 \item{neighbour_4_CellId3}{feature from histoCAT}
 \item{neighbour_4_CellId4}{feature from histoCAT}
 \item{neighbour_4_CellId5}{feature from histoCAT}
 \item{neighbour_4_CellId6}{feature from histoCAT}
 \item{neighbour_4_CellId7}{feature from histoCAT}
 \item{neighbour_4_CellId8}{feature from histoCAT}
 \item{neighbour_4_CellId9}{feature from histoCAT}
 \item{neighbour_4_CellId10}{feature from histoCAT}
 \item{Phenograph7851534969}{feature from histoCAT}
 \item{tSNE4148542692_1}{feature from histoCAT}
 \item{tSNE4148542692_2}{feature from histoCAT}
   }
}
\usage{
data(data)
}
\description{
Data set containing 1,000 cells and 73 features which include panel antibody, neighborhood computations, and phenograph clustering.
}
\keyword{datasets}
