% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperimentToHyperFrame.R
\name{.imcExperimentToPPP}
\alias{.imcExperimentToPPP}
\title{map to point pattern from imcExperiment class.}
\usage{
.imcExperimentToPPP(caseExperiment = NULL, phenotypeToUse = 1)
}
\arguments{
\item{caseExperiment}{the subset IMC experiment to cast into a point pattern}

\item{phenotypeToUse}{the cluster id to annotate the pattern}
}
\value{
imcExperiment container converted to a point pattern set
}
\description{
map to point pattern from imcExperiment class.
}
