% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{getCoordinates}
\alias{getCoordinates}
\alias{getCoordinates,imcExperiment-method}
\alias{getCoordinates<-,imcExperiment,matrix-method}
\title{finds the spatial coords, getter.}
\usage{
getCoordinates(object)

\S4method{getCoordinates}{imcExperiment}(object)

\S4method{getCoordinates}{imcExperiment,matrix}(object) <- value
}
\arguments{
\item{object}{is IMC container}

\item{value}{matrix rows cells, columns are x,y}
}
\value{
imcExperiment container

imcExperiment container

imcExperiment container
}
\description{
finds the spatial coords, getter.
}
\examples{
data(imcdata)
getCoordinates(imcdata)
data(imcdata)
getCoordinates(imcdata)
data(imcdata)
x<-getCoordinates(imcdata)
getCoordinates(imcdata)<-as.matrix(x)
}
