% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{getNeighborhood}
\alias{getNeighborhood}
\alias{getNeighborhood,imcExperiment-method}
\alias{getNeighborhood<-}
\alias{getNeighborhood<-,imcExperiment,matrix-method}
\title{finds the neighborhood information.}
\usage{
getNeighborhood(object, ...)

\S4method{getNeighborhood}{imcExperiment}(object)

getNeighborhood(object) <- value

\S4method{getNeighborhood}{imcExperiment,matrix}(object) <- value
}
\arguments{
\item{object}{is IMC container}

\item{...}{additional arguments}

\item{value}{matrix rows cells, columns are neighborhood histoCAT output}
}
\value{
imcExperiment container

imcExperiment container
data(imcdata)
getNeighborhood(imcdata)

imcExperiment container

imcExperiment container
}
\description{
finds the neighborhood information.

slow assignment for the histoCAT neighborhood data (matrix) columns are the neighbors
}
\examples{
data(imcdata)
getNeighborhood(imcdata)
data(imcdata)
x<-matrix(1,nrow=ncol(imcdata),ncol=2)
getNeighborhood(imcdata)<-x
data(imcdata)
x<-matrix(1,nrow=ncol(imcdata),ncol=2)
getNeighborhood(imcdata)<-x
}
