% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{getNetwork<-}
\alias{getNetwork<-}
\title{re-assigns the network assignment (matrix)}
\usage{
getNetwork(object) <- value
}
\arguments{
\item{object}{is IMC container}

\item{value}{data.frame rows cells, columns are phenograph network ID}
}
\value{
imcExperiment container
}
\description{
re-assigns the network assignment (matrix)
}
\examples{
data(imcdata)
x<-data.frame(ID=seq_len(ncol(imcdata)))
getNetwork(imcdata)<-x
}
