% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcdata.R
\docType{data}
\name{imcdata}
\alias{imcdata}
\title{imcdata}
\format{
A data frame of 2,452 cells and histoCAT features which are containerized into the imcExperiment
}
\usage{
data(imcdata)
}
\description{
histoCAT output containerized as IMC container. IMC S4 data set containing 2,452 cells and 44 antibody features which include panel antibody.
}
\keyword{datasets}
