% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-class.R,
%   R/imcExperiment-methods.R
\docType{class}
\name{imcExperiment-class}
\alias{imcExperiment-class}
\alias{.imcExperiment}
\alias{cellIntensity,imcExperiment-method}
\alias{cellIntensity<-,imcExperiment,matrix-method}
\alias{pData,imcExperiment-method}
\alias{getSpatial,imcExperiment-method}
\alias{spatial}
\alias{imcExperiment-method}
\alias{getSpatial<-,imcExperiment,matrix-method}
\alias{getNetwork,imcExperiment-method}
\alias{getNetwork<-,imcExperiment,matrix-method}
\title{a summarized experiment of IMC runs}
\usage{
\S4method{cellIntensity}{imcExperiment}(object)

\S4method{cellIntensity}{imcExperiment,matrix}(object) <- value

\S4method{pData}{imcExperiment}(object)

getSpatial(object)

\S4method{getSpatial}{imcExperiment}(object)

\S4method{getSpatial}{imcExperiment,matrix}(object) <- value

getNetwork(object)

\S4method{getNetwork}{imcExperiment}(object)

\S4method{getNetwork}{imcExperiment,matrix}(object) <- value
}
\arguments{
\item{object}{imcExperiment}

\item{object}{imcExperiment}
}
\description{
returns the expression of IMC data

returns the expression of IMC data

returns the phenotypic data associated with a imcExperiment object
}
\section{Slots}{

\describe{
\item{\code{spatial}}{Spatial coordinate data. x,y coordinates.}
}}

\section{Slots}{

\describe{
  \item{\code{spatial}:}{data.frame class containing x,y coordinates}
}
}

