% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-class.R,
%   R/imcExperiment-methods.R
\docType{class}
\name{imcExperiment-class}
\alias{imcExperiment-class}
\alias{.imcExperiment}
\alias{cellIntensity,imcExperiment-method}
\alias{cellIntensity}
\alias{imcExperiment-method}
\alias{cellIntensity<-,imcExperiment,matrix-method}
\alias{getCoordinates,imcExperiment-method}
\alias{getCoordinates}
\alias{getCoordinates<-,imcExperiment,matrix-method}
\alias{getNeighborhood,imcExperiment-method}
\alias{getSpatial}
\alias{getNeighborhood<-,imcExperiment,matrix-method}
\alias{spatial}
\alias{getNetwork,imcExperiment-method}
\alias{getNetwork}
\alias{getNetwork<-,imcExperiment,matrix-method}
\alias{getDistance,imcExperiment-method}
\alias{getDistance}
\alias{getDistance<-,imcExperiment,matrix-method}
\alias{getMorphology,imcExperiment-method}
\alias{getMorphology}
\alias{getMorphology<-,imcExperiment,matrix-method}
\alias{getLabel,imcExperiment-method}
\alias{getLabel}
\alias{subsetCase,imcExperiment-method}
\alias{selectCases,imcExperiment-method}
\title{a summarized experiment of IMC runs, dimensions of the spatial and intensity data are regulated.#'}
\usage{
cellIntensity(object, ...)

\S4method{cellIntensity}{imcExperiment}(object)

\S4method{cellIntensity}{imcExperiment,matrix}(object) <- value

getCoordinates(object)

\S4method{getCoordinates}{imcExperiment}(object)

\S4method{getCoordinates}{imcExperiment,matrix}(object) <- value

getNeighborhood(object)

\S4method{getNeighborhood}{imcExperiment}(object)

\S4method{getNeighborhood}{imcExperiment,matrix}(object) <- value

getNetwork(object)

\S4method{getNetwork}{imcExperiment}(object)

\S4method{getNetwork}{imcExperiment,matrix}(object) <- value

getDistance(object)

\S4method{getDistance}{imcExperiment}(object)

\S4method{getDistance}{imcExperiment,matrix}(object) <- value

getMorphology(object)

\S4method{getMorphology}{imcExperiment}(object)

\S4method{getMorphology}{imcExperiment,matrix}(object) <- value

getLabel(object)

\S4method{getLabel}{imcExperiment}(object)

subsetCase(object, value)

\S4method{subsetCase}{imcExperiment}(object, value)

selectCases(object, value)

\S4method{selectCases}{imcExperiment}(object, value)
}
\arguments{
\item{object}{IMC container}

\item{value}{this is ROIID vector}
}
\description{
a summarized experiment of IMC runs, dimensions of the spatial and intensity data are regulated.#'

finds the intensities getter.

finds the spatial coords, getter.

finds the neighborhood information.

a slot for the histoCAT neighborhood data (matrix) columns are the neighbors

finds the network information.

assigns cell cluster assignment to the container. rows are cells and column is the cluster ID

finds the distance information.

distance matrix can be stored in the distance slot for pairwise distance

finds the morphology information.

morphological features can be stored (matrix) rows are cells and columns are Area, etc.

finds the label information.

unique cell labels can be assigned (vector)

subsets the imcExperiment to a case along with all slots for a single ROI, using for distance analysis

method to subset the slots, requires colData with column "ROIID"

subsets the imcExperiment to a case along with all slots for a selected multiple ROIs.

method to subset the slots, requires colData with column "ROIID"
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{Spatial coordinate (rows are cells and columns are x,y coordinates.}

\item{\code{cellIntensity}}{a matrix of the intensity values rows are proteins and columns are cells (SCE format)}

\item{\code{neighborHood}}{a matrix (rows are cells, columns are features) of neighborhood observations from neighborhood analysis}

\item{\code{network}}{a matrix (rows are cells) and columns are for network assignment information}

\item{\code{distance}}{a matrix of rows are cells and columns are for distance measurements}

\item{\code{uniqueLabel}}{a character for each cell for unique labeling.}
}}

\section{Slots}{

\describe{
  \item{\code{coordinates}:}{matrix class containing x,y coordinates}
  \item{\code{cellIntensity}:}{matrix class containing intensities}
  \item{\code{neighborHood}:}{matrix class containing neighborhood results}
  \item{\code{network}:}{matrix class containing cell assignment cluster}
  \item{\code{distance}:}{matrix class containing cell distance observations}
  \item{\code{uniqueLabel}:}{character class containing labels}
}
}

