% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperimentToHyperFrame.R
\name{imcExperimentToHyperFrame}
\alias{imcExperimentToHyperFrame}
\title{map to point pattern from imcExperiment class.}
\usage{
imcExperimentToHyperFrame(imcExperiment = NULL, phenotypeToUse = 1)
}
\arguments{
\item{imcExperiment}{imcExperiment class}

\item{phenotypeToUse}{the network slot can often have many columns, this is the ID for the column number to use in the network slot.}
}
\description{
map to point pattern from imcExperiment class.
}
