% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcExperiment-methods.R
\name{selectCases}
\alias{selectCases}
\alias{selectCases,imcExperiment-method}
\title{subsets the imcExperiment to a case along with all slots for a selected multiple ROIs.}
\usage{
selectCases(object, value, ...)

\S4method{selectCases}{imcExperiment}(object, value)
}
\arguments{
\item{object}{IMC container}

\item{value}{this is ROIID vector}

\item{...}{additional parameters}
}
\value{
imcExperiment container of selected cases

imcExperiment container of selected cases
}
\description{
subsets the imcExperiment to a case along with all slots for a selected multiple ROIs.

method to subset the slots, requires colData with column "ROIID"
}
\examples{
data(imcdata)
myCases<-selectCases(imcdata,c("30-BM16-202_7Pre_s1_p1_r4_a4_ac","B17_350_14post_s1_p1_r5_a5_ac"))
myCases
table(colData(myCases)$ROIID)
data(imcdata)
myCases<-selectCases(imcdata,c("30-BM16-202_7Pre_s1_p1_r4_a4_ac","B17_350_14post_s1_p1_r5_a5_ac"))
myCases
table(colData(myCases)$ROIID)
}
